Nuevos tests incluidos

+++ RowDaBoat-x64barebones-d4e1c147f975/Kernel/utilities/sem.c
@@ -20,7 +20,7 @@ typedef struct pidQueue{
 }pidQueue_t;
 
 typedef struct semaphore{
-    uint16_t counter;
+    uint64_t counter;
     char name[NAME_SIZE];
     pidQueue_t blockedProcessesPidQueue;
     uint16_t dependantProcessesCount;
